requirements.txt
django==3.2.6
django-environ==0.4.5
django-cors-headers==3.8.0
itsdangerous==2.0.1
python-decouple==3.4
dj-database-url==0.5.0

# Django REST Framework
djangorestframework==3.12.4
coreapi==2.3.3
django-filter==2.4.0
drf-nested-routers==0.93.3

six==1.16.0
django-utils-six==2.0

# tests
model-bakery==1.3.2
django-coverage==1.2.4
vcrpy==4.1.1